+++ src/types/StyleIdMap.ts
@@ -1,0 +1,2 @@
+export type StyleThemeMap = Record<string, Record<string, string>>;
+export type StyleIdMap = Record<string, string>;
